Fix five correctness bugs found in code review
- input_system: `.default_value = "NULL"` (string) → `NULL` (pointer); pressing
  Enter on an empty hostname no longer tries to connect to a host named "NULL"
- custom_io: clean TCP close now returns WS_CBIO_ERR_CONN_CLOSE instead of
  WS_SOCKET_ERROR_E, matching wolfSSH's own io.c so the peer thread receives
  WS_EOF and fires SSH_EVENT_DISCONNECTED rather than SSH_EVENT_ERROR
- renderer: remove stray (uint8_t) cast on uint32_t codepoint that truncated
  code points above U+00FF
- ui_manager/ssh_manager: add show_header parameter to ui_manager_show_auth_prompt;
  the "Authentication Required" banner is now printed only once per auth exchange
  instead of before every individual keyboard-interactive prompt
- ssh_client: set SO_SNDTIMEO before TCP connect so hung connections time out
  (SSH_CONNECT_TIMEOUT_SEC); reset to zero after a successful connect

+++ sdk_apps/sshterm/components/renderer/renderer.c
@@ -221,7 +221,7 @@ void renderer_present_if_dirty(uint32_t now_ms) {
             }
             
             // Draw character
-            draw_glyph_runs((uint8_t)cell->cp, px, py, cell->fg.rgb);
+            draw_glyph_runs(cell->cp, px, py, cell->fg.rgb);
         }
     }
